+++ deployment/metrics/variables.tf
@@ -5,5 +5,5 @@ variable "VERSION" {
   description = "(Internal) The version of the KNFSD Monitoring Dashboard."
   type        = string
   nullable    = false
-  default     = "16"
+  default     = "17"
 }