+++ deployment/terraform-module-knfsd/modules/loadbalancer/main.tf
@@ -7,16 +7,16 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.64.0"
+      version = "~> 6.66.0"
     }
     dns = {
       source  = "hashicorp/dns"
-      version = "~> 3.6.0"
+      version = "~> 3.6.2"
     }
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/modules/loadbalancer/1.1.0-beta.4"
+      "knfsd-file-cache/modules/loadbalancer/1.1.0-beta.5"
     ]
   }
 }