+++ image/resources/knfsd-metrics-agent/internal/fscache/internal/metadata/generated_config.go
@@ -1489,8 +1489,3 @@ func NewDefaultMetricsBuilderConfig() MetricsBuilderConfig {
 		Metrics: DefaultMetricsConfig(),
 	}
 }
-
-// Deprecated: Use NewDefaultMetricsBuilderConfig.
-func DefaultMetricsBuilderConfig() MetricsBuilderConfig {
-	return NewDefaultMetricsBuilderConfig()
-}