+++ image/resources/knfsd-metrics-agent/internal/mounts/internal/metadata/generated_config.go
@@ -747,8 +747,3 @@ func NewDefaultMetricsBuilderConfig() MetricsBuilderConfig {
 		Metrics: DefaultMetricsConfig(),
 	}
 }
-
-// Deprecated: Use NewDefaultMetricsBuilderConfig.
-func DefaultMetricsBuilderConfig() MetricsBuilderConfig {
-	return NewDefaultMetricsBuilderConfig()
-}