+++ image/resources/knfsd-metrics-agent/internal/slab/internal/metadata/generated_config.go
@@ -121,8 +121,3 @@ func NewDefaultMetricsBuilderConfig() MetricsBuilderConfig {
 		Metrics: DefaultMetricsConfig(),
 	}
 }
-
-// Deprecated: Use NewDefaultMetricsBuilderConfig.
-func DefaultMetricsBuilderConfig() MetricsBuilderConfig {
-	return NewDefaultMetricsBuilderConfig()
-}